move pdqfactory to new file
